erp/requirements.txt
django-codenerix>=4.0.11
django-codenerix-email
django-codenerix-extensions
django-codenerix-invoicing
django-codenerix-payments
django-codenerix-products
django-codenerix-storages
django-codenerix-geodata
django-codenerix-transports
django-codenerix-corporate
django-codenerix-pos
django-codenerix-cms
django-codenerix-reviews
django-codenerix-vending
django-extensions
social-auth-app-django
ipython
mysqlclient
